requirements.txt
pyserial~=3.5
uvicorn~=0.30.6
fastapi~=0.114.0